Assembling Configuration moved to PhpUnit adapter
